+++ src/common/config/CommonConfig.cpp
@@ -67,6 +67,7 @@ xmrig::CommonConfig::CommonConfig() :
     m_adjusted(false),
     m_apiIPv6(false),
     m_apiRestricted(true),
+    m_autoSave(true),
     m_background(false),
     m_colors(true),
     m_dryRun(false),
@@ -173,7 +174,10 @@ void xmrig::CommonConfig::printVersions()
     int length = snprintf(buf, sizeof buf, "CUDA/%d.%d ", cudaVersion / 1000, cudaVersion % 100);
 #   else
     memset(buf, 0, 16);
+
+#   if !defined(XMRIG_NO_HTTPD) || !defined(XMRIG_NO_TLS)
     int length = 0;
+#   endif
 #   endif
 
 #   if !defined(XMRIG_NO_TLS) && defined(OPENSSL_VERSION_TEXT)
@@ -217,7 +221,7 @@ bool xmrig::CommonConfig::save()
     rapidjson::PrettyWriter<rapidjson::FileWriteStream> writer(os);
     doc.Accept(writer);
 
-    fclose(fp);
+    fflush(fp);
 
     uv_fs_close(uv_default_loop(), &req, fd, nullptr);
     uv_fs_req_cleanup(&req);
@@ -279,16 +283,16 @@ bool xmrig::CommonConfig::parseBoolean(int key, bool enable)
         break;
 
     case KeepAliveKey: /* --keepalive */
-        m_pools.back().setKeepAlive(enable ? Pool::kKeepAliveTimeout : 0);
+        currentPool().setKeepAlive(enable ? Pool::kKeepAliveTimeout : 0);
         break;
 
     case TlsKey: /* --tls */
-        m_pools.back().setTLS(enable);
+        currentPool().setTLS(enable);
         break;
 
 #   ifndef XMRIG_PROXY_PROJECT
     case NicehashKey: /* --nicehash */
-        m_pools.back().setNicehash(enable);
+        currentPool().setNicehash(enable);
         break;
 #   endif
 
@@ -308,8 +312,12 @@ bool xmrig::CommonConfig::parseBoolean(int key, bool enable)
         m_apiRestricted = enable;
         break;
 
-    case IConfig::DryRunKey: /* --dry-run */
+    case DryRunKey: /* --dry-run */
         m_dryRun = enable;
+        break;
+
+    case AutoSaveKey:
+        m_autoSave = enable;
         break;
 
     default:
@@ -328,13 +336,15 @@ bool xmrig::CommonConfig::parseString(int key, const char *arg)
         break;
 
     case UserpassKey: /* --userpass */
-        if (!m_pools.back().setUserpass(arg)) {
+        if (!currentPool().setUserpass(arg)) {
             return false;
         }
 
         break;
 
     case UrlKey: /* --url */
+        fixup();
+
         if (m_pools.size() > 1 || m_pools[0].isValid()) {
             Pool pool(arg);
 
@@ -353,23 +363,23 @@ bool xmrig::CommonConfig::parseString(int key, const char *arg)
         break;
 
     case UserKey: /* --user */
-        m_pools.back().setUser(arg);
+        currentPool().setUser(arg);
         break;
 
     case PasswordKey: /* --pass */
-        m_pools.back().setPassword(arg);
+        currentPool().setPassword(arg);
         break;
 
     case RigIdKey: /* --rig-id */
-        m_pools.back().setRigId(arg);
+        currentPool().setRigId(arg);
         break;
 
     case FingerprintKey: /* --tls-fingerprint */
-        m_pools.back().setFingerprint(arg);
+        currentPool().setFingerprint(arg);
         break;
 
     case VariantKey: /* --variant */
-        m_pools.back().algorithm().parseVariant(arg);
+        currentPool().algorithm().parseVariant(arg);
         break;
 
     case LogFileKey: /* --log-file */
@@ -395,7 +405,7 @@ bool xmrig::CommonConfig::parseString(int key, const char *arg)
     case RetriesKey:     /* --retries */
     case RetryPauseKey:  /* --retry-pause */
     case ApiPort:        /* --api-port */
-    case PrintTimeKey:   /* --cpu-priority */
+    case PrintTimeKey:   /* --print-time */
         return parseUint64(key, strtol(arg, nullptr, 10));
 
     case BackgroundKey: /* --background */
@@ -457,11 +467,11 @@ bool xmrig::CommonConfig::parseInt(int key, int arg)
         break;
 
     case KeepAliveKey: /* --keepalive */
-        m_pools.back().setKeepAlive(arg);
+        currentPool().setKeepAlive(arg);
         break;
 
     case VariantKey: /* --variant */
-        m_pools.back().algorithm().parseVariant(arg);
+        currentPool().algorithm().parseVariant(arg);
         break;
 
     case DonateLevelKey: /* --donate-level */
@@ -487,4 +497,31 @@ bool xmrig::CommonConfig::parseInt(int key, int arg)
     }
 
     return true;
+}
+
+
+Pool &xmrig::CommonConfig::currentPool()
+{
+    fixup();
+
+    return m_pools.back();
+}
+
+
+void xmrig::CommonConfig::fixup()
+{
+    if (m_state == NoneState) {
+        return;
+    }
+
+    if (m_pools.empty()) {
+        if (!m_activePools.empty()) {
+            std::swap(m_pools, m_activePools);
+        }
+        else {
+            m_pools.push_back(Pool());
+        }
+
+        m_state = NoneState;
+    }
 }