+++ src/core/ConfigLoader_platform.h
@@ -43,53 +43,60 @@ namespace xmrig {
 static char const usage[] = "\
 Usage: " APP_ID " [OPTIONS]\n\
 Options:\n\
-  -a, --algo=ALGO          specify the algorithm to use\n\
-                             cryptonight\n"
+  -a, --algo=ALGO              specify the algorithm to use\n\
+                                 cryptonight\n"
 #ifndef XMRIG_NO_AEON
 "\
-                             cryptonight-lite\n"
+                                 cryptonight-lite\n"
 #endif
 #ifndef XMRIG_NO_SUMO
 "\
-                             cryptonight-heavy\n"
+                                 cryptonight-heavy\n"
 #endif
 "\
-  -o, --url=URL            URL of mining server\n\
-  -O, --userpass=U:P       username:password pair for mining server\n\
-  -u, --user=USERNAME      username for mining server\n\
-  -p, --pass=PASSWORD      password for mining server\n\
-      --rig-id=ID          rig identifier for pool-side statistics (needs pool support)\n\
-  -k, --keepalive          send keepalived for prevent timeout (need pool support)\n\
-  -r, --retries=N          number of times to retry before switch to backup server (default: 5)\n\
-  -R, --retry-pause=N      time to pause between retries (default: 5)\n\
-      --opencl-devices=N   list of OpenCL devices to use.\n\
-      --opencl-launch=IxW  list of launch config, intensity and worksize\n\
-      --opencl-affinity=N  affine GPU threads to a CPU\n\
-      --opencl-platform=N  OpenCL platform index\n\
-      --opencl-loader=N    path to OpenCL-ICD-Loader (OpenCL.dll or libOpenCL.so)\n\
-      --print-platforms    print available OpenCL platforms and exit\n\
-      --no-color           disable colored output\n\
-      --variant            algorithm PoW variant\n\
-      --donate-level=N     donate level, default 5%% (5 minutes in 100 minutes)\n\
-      --user-agent         set custom user-agent string for pool\n\
-  -B, --background         run the miner in the background\n\
-  -c, --config=FILE        load a JSON-format configuration file\n\
-  -l, --log-file=FILE      log all output to a file\n"
+  -o, --url=URL                URL of mining server\n\
+  -O, --userpass=U:P           username:password pair for mining server\n\
+  -u, --user=USERNAME          username for mining server\n\
+  -p, --pass=PASSWORD          password for mining server\n\
+      --rig-id=ID              rig identifier for pool-side statistics (needs pool support)\n\
+  -k, --keepalive              send keepalived for prevent timeout (needs pool support)\n\
+      --nicehash               enable nicehash.com support\n\
+      --tls                    enable SSL/TLS support (needs pool support)\n\
+      --tls-fingerprint=F      pool TLS certificate fingerprint, if set enable strict certificate pinning\n\
+  -r, --retries=N              number of times to retry before switch to backup server (default: 5)\n\
+  -R, --retry-pause=N          time to pause between retries (default: 5)\n\
+      --opencl-devices=N       list of OpenCL devices to use.\n\
+      --opencl-launch=IxW      list of launch config, intensity and worksize\n\
+      --opencl-strided-index=N list of strided_index option values for each thread\n\
+      --opencl-mem-chunk=N     list of mem_chunk option values for each thread\n\
+      --opencl-comp-mode=N     list of comp_mode option values for each thread\n\
+      --opencl-affinity=N      list of affinity GPU threads to a CPU\n\
+      --opencl-platform=N      OpenCL platform index\n\
+      --opencl-loader=N        path to OpenCL-ICD-Loader (OpenCL.dll or libOpenCL.so)\n\
+      --print-platforms        print available OpenCL platforms and exit\n\
+      --no-cache               disable OpenCL cache\n\
+      --no-color               disable colored output\n\
+      --variant                algorithm PoW variant\n\
+      --donate-level=N         donate level, default 5%% (5 minutes in 100 minutes)\n\
+      --user-agent             set custom user-agent string for pool\n\
+  -B, --background             run the miner in the background\n\
+  -c, --config=FILE            load a JSON-format configuration file\n\
+  -l, --log-file=FILE          log all output to a file\n"
 # ifdef HAVE_SYSLOG_H
 "\
-  -S, --syslog             use system log for output messages\n"
+  -S, --syslog                 use system log for output messages\n"
 # endif
 "\
-      --nicehash           enable nicehash/xmrig-proxy support\n\
-      --print-time=N       print hashrate report every N seconds\n\
-      --api-port=N         port for the miner API\n\
-      --api-access-token=T access token for API\n\
-      --api-worker-id=ID   custom worker-id for API\n\
-      --api-id=ID          custom instance ID for API\n\
-      --api-ipv6           enable IPv6 support for API\n\
-      --api-no-restricted  enable full remote access (only if API token set)\n\
-  -h, --help               display this help and exit\n\
-  -V, --version            output version information and exit\n\
+      --print-time=N           print hashrate report every N seconds\n\
+      --api-port=N             port for the miner API\n\
+      --api-access-token=T     access token for API\n\
+      --api-worker-id=ID       custom worker-id for API\n\
+      --api-id=ID              custom instance ID for API\n\
+      --api-ipv6               enable IPv6 support for API\n\
+      --api-no-restricted      enable full remote access (only if API token set)\n\
+      --dry-run                test configuration and exit\n\
+  -h, --help                   display this help and exit\n\
+  -V, --version                output version information and exit\n\
 ";
 
 
@@ -131,7 +138,7 @@ static struct option const options[] = {
     { "opencl-devices",       1, nullptr, xmrig::IConfig::OclDevicesKey     },
     { "opencl-launch",        1, nullptr, xmrig::IConfig::OclLaunchKey      },
     { "opencl-platform",      1, nullptr, xmrig::IConfig::OclPlatformKey    },
-    { "opencl-srided-index",  1, nullptr, xmrig::IConfig::OclSridedIndexKey },
+    { "opencl-strided-index", 1, nullptr, xmrig::IConfig::OclSridedIndexKey },
     { "opencl-mem-chunk",     1, nullptr, xmrig::IConfig::OclMemChunkKey    },
     { "opencl-unroll-factor", 1, nullptr, xmrig::IConfig::OclUnrollKey      },
     { "opencl-unroll",        1, nullptr, xmrig::IConfig::OclUnrollKey      },
@@ -158,6 +165,7 @@ static struct option const config_options[] = {
     { "opencl-platform",   1, nullptr, xmrig::IConfig::OclPlatformKey },
     { "cache",             0, nullptr, xmrig::IConfig::OclCacheKey    },
     { "opencl-loader",     1, nullptr, xmrig::IConfig::OclLoaderKey   },
+    { "autosave",          0, nullptr, xmrig::IConfig::AutoSaveKey    },
     { nullptr,             0, nullptr, 0 }
 };