+++ src/core/Controller.cpp
@@ -27,6 +27,7 @@
 
 #include "amd/OclLib.h"
 #include "common/config/ConfigLoader.h"
+#include "common/cpu/Cpu.h"
 #include "common/interfaces/IControllerListener.h"
 #include "common/log/ConsoleLog.h"
 #include "common/log/FileLog.h"
@@ -34,7 +35,6 @@
 #include "common/Platform.h"
 #include "core/Config.h"
 #include "core/Controller.h"
-#include "Cpu.h"
 #include "net/Network.h"
 
 
@@ -76,6 +76,12 @@ xmrig::Controller::~Controller()
     ConfigLoader::release();
 
     delete d_ptr;
+}
+
+
+bool xmrig::Controller::isDone() const
+{
+    return ConfigLoader::isDone();
 }